# Concurrency optimization for native graph loading