… CHANGELOG.md
… src/test/java/org/opensearch/knn/index/memory/NativeMemoryEntryContextTests.java
… src/test/java/org/opensearch/knn/index/memory/NativeMemoryLoadStrategyTests.java