@0ctopus13prime 2024-12-20
HI @Gankris96, thank you for the PR. 
I see this clearly will bring benefit for the cases where multiple threads are competing each other.
But just curious, after this fix, how much performance gain do you see?