@Gankris96 2024-12-20
> HI @Gankris96, thank you for the PR. I see this clearly will bring benefit for the cases where multiple threads are competing each other. But just curious, after this fix, how much performance gain do you see?

Hi @0ctopus13prime  - yes i am working on getting the benchmarking numbers. Primarily trying to test it on remote store backed index to see the performance gains.