@Gankris96 2025-01-07
@0ctopus13prime Have added the benchmarking numbers above and here is the summary -

1. Regular index testing with warmup
```
|                  Percentile | Without Fix | With Fix | Delta (ms) | Improvement |
|----------------------------|-------------|----------|------------|-------------|
|                       50th |     23.2200 | 17.1917 |    -6.0283 |     -26.0% |
|                       90th |     61.1095 | 32.2796 |   -28.8299 |     -47.2% |
|                       99th |    378.3850 | 285.9520|   -92.4330 |     -24.4% |
|                      100th |    378.4050 | 287.9980|   -90.4070 |     -23.9% |
```
2. Regular index testing with clear cache after each query 
```
| Query #   | Without Fix | With Fix  | Delta (ms) | Improvement |
|-----------|-------------|-----------|------------|-------------|
| Query 1   |    3069.56 |  2041.92  | -1027.64   |    -33.5%   |
| Query 2   |    2761.68 |  1222.25  | -1539.43   |    -55.7%   |
| Query 3   |    1722.23 |   878.90  |  -843.33   |    -49.0%   |
| Query 4   |    1502.61 |   864.43  |  -638.18   |    -42.5%   |
| Query 5   |    1691.39 |   860.18  |  -831.21   |    -49.1%   |
| Query 6   |    1705.31 |   872.60  |  -832.71   |    -48.8%   |
| Query 7   |    2861.13 |  1028.30  | -1832.83   |    -64.1%   |
| Query 8   |    1019.70 |  1131.18  |  +111.48   |    +10.9%   |
| Query 9   |    1194.15 |   883.22  |  -310.94   |    -26.0%   |
| Query 10* |        N/A |   850.83  |      N/A   |      N/A    |


Key Statistics:
- Average improvement (excluding Query 10): -39.8%
- Average time without fix: 1947.53 ms
- Average time with fix: 1063.38 ms
```

--- 
3. Remote store index with warmup 
```
|                  Percentile | Without Fix | With Fix | Delta (ms) | Improvement |
|----------------------------|-------------|----------|------------|-------------|
|                       50th |     22.3057 | 18.5364  |    -3.7693 |     -16.9% |
|                       90th |     86.1489 | 42.2512  |   -43.8977 |     -51.0% |
|                       99th |    708.6770 | 583.2010 |  -125.4760 |     -17.7% |
|                      100th |    709.3910 | 583.9680 |  -125.4230 |     -17.7% |
```
4. Remote store index with clear cache after each - 
```
| Query #   | Without Fix | With Fix  | Delta (ms) | Improvement |
|-----------|-------------|-----------|------------|-------------|
| Query 1   |    2381.68 |   886.21  | -1495.47   |    -62.8%   |
| Query 2   |     933.66 |   945.63  |   +11.97   |     +1.3%   |
| Query 3   |     984.45 |   951.20  |   -33.25   |     -3.4%   |
| Query 4   |     977.13 |  1037.99  |   +60.86   |     +6.2%   |
| Query 5   |     877.56 |   982.43  |  +104.87   |    +11.9%   |
| Query 6   |     984.30 |  1116.40  |  +132.10   |    +13.4%   |
| Query 7   |    1082.44 |  1152.06  |   +69.62   |     +6.4%   |
| Query 8   |    1118.12 |   985.59  |  -132.53   |    -11.9%   |
| Query 9   |     921.84 |   949.89  |   +28.05   |     +3.0%   |
| Query 10  |     902.73 |  1006.75  |  +104.02   |    +11.5%   |

Key Statistics:
- Average time without fix: 1116.39 ms
- Average time with fix: 1001.41 ms
- Overall average improvement: -2.4%
```

Overall we dont see much improvement in the remote-store case but with a more constrained environment (larger graphs) and more concurrent searches, we should see even more improvement.

@0ctopus13prime , @kotwanikunal , @shatejas  let me know your thoughts