@Gankris96 2025-01-08
Here is the benchmarking comparison with concurrent segment search enabled on remote-store index 

```
 Query #   | Without Fix | With Fix  | Delta (ms) | Improvement |
|-----------|-------------|-----------|------------|-------------|
| Query 1   |    1940.73 |  1899.53  |   -41.20   |     -2.1%   |
| Query 2   |     915.54 |   957.23  |   +41.69   |     +4.6%   |
| Query 3   |     883.05 |   874.13  |    -8.92   |     -1.0%   |
| Query 4   |     850.17 |   859.59  |    +9.42   |     +1.1%   |
| Query 5   |     936.52 |   748.44  |  -188.08   |    -20.1%   |
| Query 6   |     857.33 |   841.82  |   -15.51   |     -1.8%   |
| Query 7   |     870.06 |   773.51  |   -96.55   |    -11.1%   |
| Query 8   |     825.31 |   847.89  |   +22.58   |     +2.7%   |
| Query 9   |     841.98 |   897.34  |   +55.37   |     +6.6%   |
| Query 10  |     828.68 |   860.58  |   +31.90   |     +3.9%   |

Key Statistics:
- Average time without fix: 974.93 ms
- Average time with fix: 956.00 ms
```

--- 

Searchable snapshot indices restored_target_index and restored_target_index2
```
red    open target_index           01BL4u79TYSSitkUCktcVg 1 1                      
yellow open restored_target_index  un9f_mzTRQKAQWniE0bqoA 1 1 1000000 0 1.4gb 1.4gb
red    open target_index2          GuFw20DbQ067yZRtIy34ow 1 1                      
yellow open restored_target_index2 S8lTIANoT1ytZ4jVJsxgAQ 1 1 1000000 0 1.4gb 1.4gb
red    open million_songs          -rbMHk1QQF2U7WlcMgO9Dw 1 1                      


sh-4.2$ curl "localhost:9200/restored*/_settings?pretty"
{
  "restored_target_index" : {
    "settings" : {
      "index" : {
        "replication" : {
          "type" : "SEGMENT"
        },
        "knn.advanced" : {
          "approximate_threshold" : "20"
        },
        "number_of_shards" : "1",
        "remote_store" : {
          "translog" : {
            "repository" : "local-repo"
          },
          "enabled" : "true",
          "segment" : {
            "repository" : "local-repo"
          }
        },
        "provided_name" : "target_index",
        "searchable_snapshot" : {
          "index" : {
            "id" : "nNNDvxFeQ1qFJl4LjcrBkQ"
          },
          "shard_path_type" : "FIXED",
          "repository" : "my_backup",
          "snapshot_id" : {
            "name" : "snapshot_1",
            "uuid" : "SnNo_A97S_yoUhfDOn8qZA"
          }
        },
        "knn" : "true",
        "creation_date" : "1736203092702",
        "store" : {
          "type" : "remote_snapshot"
        },
        "number_of_replicas" : "1",
        "uuid" : "un9f_mzTRQKAQWniE0bqoA",
        "version" : {
          "created" : "137217827"
        }
      }
    }
  },
  "restored_target_index2" : {
    "settings" : {
      "index" : {
        "replication" : {
          "type" : "SEGMENT"
        },
        "number_of_shards" : "1",
        "remote_store" : {
          "translog" : {
            "repository" : "local-repo"
          },
          "enabled" : "true",
          "segment" : {
            "repository" : "local-repo"
          }
        },
        "provided_name" : "target_index2",
        "searchable_snapshot" : {
          "index" : {
            "id" : "4fkcDpVGSeuHpyKBlJEHWg"
          },
          "shard_path_type" : "FIXED",
          "repository" : "my_backup",
          "snapshot_id" : {
            "name" : "snapshot_2",
            "uuid" : "AAQ6KfAsTGmHKdvTxXUCTw"
          }
        },
        "knn" : "true",
        "creation_date" : "1736203842991",
        "store" : {
          "type" : "remote_snapshot"
        },
        "number_of_replicas" : "1",
        "uuid" : "S8lTIANoT1ytZ4jVJsxgAQ",
        "version" : {
          "created" : "137217827"
        }
      }
    }
  }
}
sh-4.2$ 
```

### Cluster settings 

```
{
    "persistent": {
        "cluster": {
            "blocks": {
                "create_index": "true"
            }
        },
        "knn": {
            "feature": {
                "cache": {
                    "force_evict": {
                        "enabled": "true"
                    }
                }
            }
        },
        "search": {
            "concurrent_segment_search": {
                "mode": "all"
            }
        }
    },
    "transient": {}
}
```

## Searchable snapshots - 
```
| Query #   | Without Fix | With Fix  | Delta (ms) | Improvement |
|-----------|-------------|-----------|------------|-------------|
| Query 1   |    3127.25 |   108.58  | -3018.67   |    -96.5%   |
| Query 2   |     119.83 |   101.21  |   -18.62   |    -15.5%   |
| Query 3   |     112.18 |    96.32  |   -15.86   |    -14.1%   |
| Query 4   |     125.90 |   101.54  |   -24.36   |    -19.3%   |
| Query 5   |     118.81 |   102.40  |   -16.41   |    -13.8%   |
| Query 6   |     122.63 |   105.12  |   -17.51   |    -14.3%   |
| Query 7   |     131.87 |   110.51  |   -21.36   |    -16.2%   |
| Query 8   |     112.92 |   107.89  |    -5.03   |     -4.5%   |
| Query 9   |     112.51 |   102.68  |    -9.83   |     -8.7%   |

Key Statistics:
- Average time without fix: 453.77 ms
- Average time with fix: 104.03 ms
- Overall average improvement: -77.1%

```

The searchable snapshot case shows improvement @navneet1v